* next: (1828 commits)
  Use MPI_COMM_NULL rather than nullptr
  Apply clang-format changes
  Apply clang-format changes
  Don't compile LaplaceXY2 with 3D metrics
  Apply clang-format changes
  Apply clang-format changes
  Laplacexy2 modifications
  Apply clang-format changes
  Apply clang-format changes
  nvcc needs toString(HYPRE_SOLVER_TYPE)
  nvcc now seems to need the factory flags
  Update fmt to latest version
  Apply clang-format changes
  Output BoutException.what() in PhysicsModel main()
  use hasParallelSlices
  Apply clang-format changes
  Only destroy if it exists
  Backgrounds don't work with parallel BCs
  Ensure that parallel slices exist
  CI: Remove unused stage from clang-tidy-review job
  ...

+++ src/sys/bout_types.cxx
@@ -9,7 +9,7 @@ const std::string& safeAt(const std::map<T, std::string>& mymap, T t) {
   AUTO_TRACE();
   auto found = mymap.find(t);
   if (found == mymap.end()) {
-    throw BoutException("Did not find enum {:d}", t);
+    throw BoutException("Did not find enum {:d}", static_cast<int>(t));
   }
   return found->second;
 }
